requirements.txt
asgiref==3.7.2
Django==4.2.7
django-colorfield==0.10.1
Pillow==10.1.0
python-slugify==8.0.1
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.8.0
